+++ src/components/MDSFail/index.tsx
@@ -22,7 +22,7 @@ export const MDSFail: React.FC<PropsWithChildren> = () => {
     if (IS_MINIMA_BROWSER) {
       return Android.closeWindow();
     }
-  }
+  };
 
   if (hasUpdated) {
     return <div />;
@@ -35,25 +35,26 @@ export const MDSFail: React.FC<PropsWithChildren> = () => {
           {display && (
             <div
               onClick={(evt) => evt.stopPropagation()}
-              className="mx-auto absolute w-full h-full z-[60] flex items-center justify-center text-black"
+              className="mx-auto absolute w-full h-full z-[1000] flex items-center justify-center text-black"
             >
               <div className="relative z-[80] w-full max-w-[420px] px-5">
                 <animated.div style={style}>
                   <div className="modal text-center text-white core-black-contrast-2 box-shadow-lg rounded p-6 lg:p-8 mb-6 mx-auto relative overflow-hidden">
                     <h5 className="text-xl -mt-1 mb-4 font-bold">Connection failed</h5>
                     {!isMinimaBrowser && (
-                      <p className="mb-6">
-                        Your session is invalid, please click the button below to log in again.
-                      </p>
+                      <p className="mb-6">Your session is invalid, please click the button below to log in again.</p>
                     )}
                     {isMinimaBrowser && (
                       <p className="mb-6">
-                        Your session is invalid, press the button below to open the title bar, then press the three vertical dots and select refresh from the menu.
+                        Your session is invalid, press the button below to open the title bar, then press the three
+                        vertical dots and select refresh from the menu.
                       </p>
                     )}
                     <div className="flex flex-col gap-4">
                       {!isMinimaBrowser && (
-                        <Button onClick={goToLoginPage} variant="secondary">Login</Button>
+                        <Button onClick={goToLoginPage} variant="secondary">
+                          Login
+                        </Button>
                       )}
                       {isMinimaBrowser && (
                         <>